requirements.txt
aiogram~=2.12
environs~=8.0.0
asyncpg
python-dotenv
PyMySQL~=1.0.2
mysql-connector-python~=8.0.24